requirements.txt
pip==24.3.1
pytest==8.3.3
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-order==1.3.0
selenium>=4.22.0,<4.25.0
allure-pytest==2.13.5
allure-python-commons==2.13.5